requirements.txt
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.7.22
click==8.1.7
dnspython==2.4.2
email-validator==2.1.0.post1
exceptiongroup==1.1.3
fastapi==0.104.1
fishauth==0.1.4
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.1
httptools==0.6.1
httpx==0.25.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
orjson==3.9.10
psycopg2-binary==2.9.9
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.8.0
ujson==5.8.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
